requirements.txt
aws-cdk-lib
constructs
boto3
pandas
requests
aws-cdk.core
